+++ apps/web/modules/import/index.ts
@@ -1,0 +1,5 @@
+import { importCSVFile } from './import';
+
+export const Import = {
+  importCSVFile,
+};
